+++ src/metadata/metadata.composable.ts
@@ -1,6 +1,7 @@
 import api from "@/api/api";
 import { useResourceStore } from "@/store/resource.store";
 import useSpin from "@/spin/spin.composable";
+import type { Metadata } from "@/store/resource.types";
 
 /** Tracks fully loaded resources, so subsequent load calls can be skipped. */
 const isFresh: Record<string, true> = {};
@@ -10,20 +11,22 @@ export default function useMetadata(resourceId: string) {
   const { spin } = useSpin();
 
   /** Load data about a metadata and store it. */
-  async function loadMetadata(): Promise<void> {
+  async function loadMetadata(): Promise<Metadata | undefined> {
     // Make sure the resource has an entry in the store.
-    await resourceStore.loadResource(resourceId);
+    const resource = (await resourceStore.loadResource(resourceId)) as Metadata;
 
     // Skip if already loaded.
-    if (isFresh[resourceId]) return;
-
-    // Load remaining essential info, unless removed.
-    if (resourceId in resourceStore.metadatas) {
-      await loadYaml();
+    if (!isFresh[resourceId]) {
+      // Load remaining essential info, unless removed.
+      if (resourceId in resourceStore.metadatas) {
+        await loadYaml();
+      }
     }
 
     // Remember to skip loading next time.
     isFresh[resourceId] = true;
+
+    return resource;
   }
 
   /** Load and store the metadata yaml string. */