+++ src/resource/ResourceRedirectView.vue
@@ -4,11 +4,13 @@ import useResourceIdParam from "@/resource/resourceIdParam.composable";
 import { useResourceStore } from "@/store/resource.store";
 import SpinIndicator from "@/spin/SpinIndicator.vue";
 import useMessenger from "@/message/messenger.composable";
+import useNotFound from "@/components/notfound.composable";
 
 const router = useRouter();
 const resourceId = useResourceIdParam();
 const resourceStore = useResourceStore();
 const { alertError } = useMessenger();
+const { handle404 } = useNotFound();
 
 // Load resource and immediately redirect according to its type.
 resourceStore
@@ -17,6 +19,7 @@ resourceStore
     if (resource) router.push(`/library/${resource.type}/${resourceId}`);
     else router.push("/library");
   })
+  .catch(handle404)
   .catch(alertError);
 </script>